Moved environment tests under own directory, test environment enum label
